Fixed commands to no-op when stack is too small
According to project spec, when the stack is too small to execute a
command, nothing happens. I've interpreted this to mean that if the
stack is size one, but two pops are required, the stack keeps its one
element. However, this may be incorrect.

Also, I fixed the pop ordering for subtract, modulo, and divide